New_Proj/requirements.txt

# Django Backend
Django>=4.2,<5.1
djangorestframework>=3.14
python-dotenv>=1.0


# Automation

selenium>=4.15
requests>=2.31
webdriver-manager>=4.0


# Server / Stability

gunicorn>=21.2
urllib3>=2.0
certifi>=2024.2
